multimodal_coercion/requirements.txt
numpy>=1.26
scipy>=1.11
pandas>=2.1
opencv-python>=4.8
mediapipe>=0.10
tensorflow>=2.14
torch>=2.1
torchvision>=0.16
transformers>=4.36
sentencepiece>=0.1.99
tokenizers>=0.15
openai-whisper>=20231117
librosa>=0.10
soundfile>=0.12
streamlit>=1.29
scikit-learn>=1.3
pyyaml>=6.0
matplotlib>=3.8
seaborn>=0.13
joblib>=1.3
fastapi>=0.110
uvicorn>=0.24
python-multipart>=0.0.9
ffmpeg-python>=0.2.0
